+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # dag
 
+## 0.1.3
+
+### Patch Changes
+
+- add fn and variable to save the package names in runner
+
+## 0.1.2
+
+### Patch Changes
+
+- add find_pkg_json function
+
 ## 0.1.1
 
 ### Patch Changes